backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9 
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9 
python-decouple==3.8
pydantic[email]==2.5.0
pydantic-settings==2.1.0
numpy<2.0
shapely==2.0.2
faker==20.1.0
schedule==1.2.1